// src/app/interfaces/Burger.ts
export interface Burger{

    id?: number;
    nombre_hamburguesa: string;
    tipo_carne: string;
    carne: number;
    lechuga: boolean;
    cebolla: boolean;
    aros_cebolla: boolean;
    tomate: boolean;
    tocino: boolean;
    huevo: boolean;
    jamon: boolean;
    jalapeno: boolean;
    pepinillo: boolean;
    queso_cheddar: boolean;
    champinones: boolean;
    nachos: boolean;
    salsa: string; //ketchup, mostaza, jack daniel, bluecheese, guacamole, buffalo, barbacoa, nada
}